perbaikan model dengan menggunakan int64

+++ models/group.go
@@ -31,7 +31,7 @@ type Group struct {
 	Description string `gorm:"type:text" json:"group_description"`
 
 	OwnerID     uuid.UUID `gorm:"type:uuid;not null" json:"group_owner_id"`
-	MemberCount int64     `gorm:"-:migration;->" json:"member_count"`
+	MemberCount int       `gorm:"-:migration;->" json:"member_count"`
 	// WalletID uuid.UUID `gorm:"type:uuid;not null" json:"group_wallet_id"`
 
 	Members []GroupMember `gorm:"foreignKey:GroupID" json:"members"`
